started logging things

+++ src/entities/admin/admin.py
@@ -1,4 +1,3 @@
-from sys import exception
 from ..employee.employee import Employee
 from ...utils.async_pg_db_utils import (
     read_fields_from_record,
@@ -18,10 +17,8 @@ import json
 
 
 class Admin(Employee):
-    # def __init__(self, employee_info):
-    #     super().__init__((*employee_info, "admin"))
-    def __init__(self):
-        pass
+    def __init__(self, emp_id,logger):
+        super().__init__(emp_id,logger)
 
     async def get_req(self, status_list):
         data = await read_fields_from_record(
@@ -54,36 +51,35 @@ class Admin(Employee):
 
     async def commit_request(self, req_id):
         # get request information
-        print(req_id)
         request = await read_fields_from_record("requests", "*", "request_id", [req_id])
-        request = parse_requests(request)
-        request = request[0]
-        if not request:
+        if not len(request) == 1:
             raise ValueError("Request does not exist")
-        if request["status"] == "approved_by_hr":
-            # parse the json obj of update to updated dict
-            updated_info = json.loads(request["updated_info"])
-
-            # the address field has sub fields and should be stored as a stringified json obj
-            updated_info["address"] = json.dumps(updated_info["address"])
-
-            # update the employee record
-            await update_one_record(
-                "employees", updated_info, "empId", request["created_by"]
-            )
-
-            # update the request status to committed and add commit time
-            request["update_committed_at"] = ceil(time.time())
-            request["request_status"] = "committed"
-            await update_one_record("requests", request, "request_id", req_id)
-            return True
         else:
-            raise ValueError("Request can't be committed")
+            request = parse_requests(request)
+            request = request[0]
+            if request["status"] == "approved_by_hr":
+                # parse the json obj of update to updated dict
+                updated_info = json.loads(request["updated_info"])
+
+                # the address field has sub fields and should be stored as a stringified json obj
+                updated_info["address"] = json.dumps(updated_info["address"])
+
+                # update the employee record
+                await update_one_record(
+                    "employees", updated_info, "empId", request["created_by"]
+                )
+
+                # update the request status to committed and add commit time
+                request["update_committed_at"] = ceil(time.time())
+                request["request_status"] = "committed"
+                await update_one_record("requests", request, "request_id", req_id)
+                return True
+            else:
+                raise ValueError("Request can't be committed")
 
     async def create_new_employee(self, new_employee):
         try:
             new_employee["password"] = hash_pass(new_employee["password"])
-            new_employee["address"] = json.dumps(new_employee["address"])
             if await check_if_exists_in_db("employees", "email", new_employee["email"]):
                 raise ValueError(
                     f"User with email {new_employee['email']} already exists"
@@ -99,7 +95,7 @@ class Admin(Employee):
     async def create_new_relation(self, emp_id, reports_to_emp_id):
         try:
             if reports_to_emp_id == 0 or await check_if_exists_in_db(
-                "employees", "empId", reports_to_emp_id
+                    "employees", "empId", reports_to_emp_id
             ):
                 new_relation = {}
                 new_relation["employee"] = emp_id
@@ -117,12 +113,12 @@ class Admin(Employee):
                 return False
             else:
                 if (
-                    len(
-                        await read_by_multiple_attributes(
-                            "employees", "*", ["user_type", "empId"], ["admin", emp_id]
+                        len(
+                            await read_by_multiple_attributes(
+                                "employees", "*", ["user_type", "empId"], ["admin", emp_id]
+                            )
                         )
-                    )
-                    == 1
+                        == 1
                 ):
                     raise ValueError("Deleting last admin not allowed")
                 reporting_employees = await read_fields_from_record(
@@ -152,6 +148,8 @@ class Admin(Employee):
             raise err
 
     async def delete_employee(self, emp_id):
+        if not await check_if_exists_in_db("employees", "empId", emp_id):
+            raise ValueError("Employee does not exist.")
         return await delete_from_table("employees", "empId", emp_id)
 
     async def get_reports_to(self, emp_id):
@@ -170,9 +168,8 @@ class Admin(Employee):
         try:
             if await check_if_exists_in_db("employees", "empid", emp_id):
                 if reports_to_emp_id and await check_if_exists_in_db(
-                    "employees", "empId", reports_to_emp_id
+                        "employees", "empId", reports_to_emp_id
                 ):
-                    print("in normal update")
                     return await update_one_record(
                         "relations",
                         {"reports_to": reports_to_emp_id},
